Wrong track selected if unplayable tracks are in a browsed view
Noticed this with SoundCloud, but if there are unplayable tracks (some tracks appear in the liked list but turn out to be 404 errors) in one of the browsed views, then they will be removed from the tracklist in Mopidy. This means that the code to play the selected track from the list selects the wrong track as the index doesn't get updated.

The code I wrote uses trackid to select the track, rather than incrementing thru a tracklist.